controller: use default namespace for event recorder
Prior to this change, we were creating a event recorder on a given
namespace. When registering events for node, the following would happen:

Unable to write event: 'can't create an event with namespace 'default'
in namespace 'kube-system'' (may retry after sleeping)

This change modifies the event recorder to be created using the default
namespace, which will allow events to be recorder on that namespace.